# api/openapi.yaml
openapi: 3.0.3
info:
  title: Lux Trading API
  description: |
    One surface for every market a Lux interface shows.

    A market on a Lux chain is quoted from our own contracts. A market on any
    other chain is quoted from the pools Uniswap deployed there, read directly
    over that chain's JSON-RPC — no vendor API, no upstream account, nothing
    that stops working when somebody says no. The response is the same shape
    either way, because a client should not have to know which arm answered it.

    Every path lives under /v1/trade. api.lux.cloud carries the chain at
    /v1/chain, the C-chain at the root and other services besides, so this is
    one namespace with one owner rather than a handful of generic names on a
    shared host.

    Reads are public and unauthenticated. Nothing here signs or sends a
    transaction: the swap, approval and position paths return unsigned calldata
    for a wallet to sign.
  version: 3.0.0
  contact:
    name: Lux Industries
    url: https://lux.exchange
servers:
  - url: https://api.lux.cloud
    description: Production
  - url: http://localhost:8080
    description: A local gateway
...
paths:
...
  /v1/trade/positions:
    get:
      operationId: positions
      tags: [liquidity]
      summary: Liquidity positions held by an address
      parameters:
        - name: chainId
          in: query
          required: true
          schema: { type: integer, format: int64 }
        - name: owner
          in: query
          required: true
          schema: { type: string }
        - name: poolId
          in: query
          schema: { type: string }
      responses:
        '200':
          description: The positions held.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Envelope' }
...
components:
...
  schemas:
    Envelope:
      type: object
      description: Every answer carries the same envelope.
      required: [success]
      properties:
        success: { type: boolean }
        data: {}
        error:
          type: string
          description: Present only when success is false.
        requestId: { type: string }
...